Feature/it 03

+++ FridgeApp.Application/Commands/AddDefaultQuantityToMissingFridgeProducts.cs
@@ -3,6 +3,6 @@ using FridgeApp.Shared.Abstractions.Commands;
 
 namespace FridgeApp.Application.Commands
 {
-    public record AddDefaultQuantityToMissingFridgeProducts(Guid FridgeId, Guid ProductId, ushort Quantity)
+    public record AddDefaultQuantityToMissingFridgeProducts(Guid FridgeId, Guid ProductId, int Quantity)
         : ICommand;
 }
