+++ FridgeApp.Application/Commands/AddFridgeProduct.cs
@@ -3,5 +3,5 @@ using FridgeApp.Shared.Abstractions.Commands;
 
 namespace FridgeApp.Application.Commands
 {
-    public record AddFridgeProduct(Guid FridgeId, Guid ProductId, ushort Quantity) : ICommand;
+    public record AddFridgeProduct(Guid FridgeId, Guid ProductId, int Quantity) : ICommand;
 }
